requirements.txt
python-keycloak>=0.9.0
requests
python-jose
simplejson
